// src/types/components.ts
export interface FormValues {
  _id?: string;
  email?: string;
  address_name: string;
  address: string;
  city: string;
  state: string;
  zip: string;
  country: string;
  phone: string;
}

export interface ExtendedFormValues extends FormValues {
  billing: FormValues;
  shipping: FormValues;
  sameAsBilling: boolean;
}